requirements.txt
google-cloud-aiplatform
kfp
scikit-learn